Add new clan, BittenByAKitten, to the support scripts.
Yes, I really should move the clan name list into the database and add
a support script to scrape the site for them and a trigger for
managing the abbreviations.  And another trigger to set which clan is
"ours".  But that's what I _should_ do in the future.  For now, this.