Sort !alarms list chronologically (by alarmdate), and show the date if appropriate.
The date was always supposed to show, but it did not because of a
reversed inequality test in friendlytime().

Sorting the alarms chronologically by alarmdate is a new feature.
Previously they just came out in database order, which is fine if
you have a small number of alarms and mostly set them in order, but
it gets weird when you have a mix of alarms that were set well in
advance and ones that are set just a few hours before the event.
This is better.