Fix UTC reporting in friendlytime()
The reporting of UTC time in friendlytime(), when the timezone was set
to anything else, was buggy in two ways.  First, it didn't always
happen, and second, it always reported the _current_ UTC time,
regardless of whether the timezoned time it was reporting was the
current time or some other time (the latter case being relevant when
reporting for !alarms).

This commit fixes both bugs.