Flood protection in say()
Fix say() so that it prevents flooding.  There are some legitimate
situations (e.g. after fixing a filter and doing a !biff reset) when
you actually do want the bot to tell you a bunch of stuff in a row.
But if this is done with no delays, the network will kick the bot for
excessive flood.  Inserting some delays protects against that.